feat: add export subcommand
